isl_multi_templ.c: extract out isl_multi_move_dims_templ.c
This makes the function isl_multi_*_move_dims opt-in rather than opt-out.
